chore(col): rename column to col
