+++ src/index.ts
@@ -67,7 +67,7 @@ import { ClickBlock } from './util/click-block';
 import { Content } from './components/content/content';
 import { DateTime } from './components/datetime/datetime';
 import { FabContainer, FabButton, FabList } from './components/fab/fab';
-import { Col } from './components/grid/column';
+import { Col } from './components/grid/col';
 import { Grid } from './components/grid/grid';
 import { Row } from './components/grid/row';
 import { Icon } from './components/icon/icon';
@@ -140,7 +140,7 @@ export { ClickBlock } from './util/click-block';
 export { Content, ScrollEvent } from './components/content/content';
 export { DateTime } from './components/datetime/datetime';
 export { FabContainer, FabButton, FabList } from './components/fab/fab';
-export { Col } from './components/grid/column';
+export { Col } from './components/grid/col';
 export { Grid } from './components/grid/grid';
 export { Row } from './components/grid/row';
 export { Ion } from './components/ion';
